python/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio>=3.4.3",
    "mcp>=1.9.1",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "prompt-toolkit>=3.0.51",
    "pyinstaller>=6.13.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tiktoken>=0.9.0",
    "typer>=0.15.3",
    "websockets>=15.0.1",
]
